requirements.txt
openai==1.13.3
langchain==0.1.10
streamlit==1.31.1
python-dotenv
langchain-openai==0.0.8
PyPDF2==3.0.1
faiss-cpu==1.8.0
tiktoken==0.6.0
langchain-community==0.0.25